Remove NetworkDrivesHandler
